refactor: WhereBuilder 에게 clause 만드는 책임
