+++ src/main/java/persistence/sql/dml/DMLQueryGenerator.java
@@ -2,12 +2,10 @@ package persistence.sql.dml;
 
 import persistence.sql.mapping.*;
 
-import java.util.List;
 import java.util.stream.Collectors;
 
 public class DMLQueryGenerator {
     private final TableData table;
-    private final DataTypeMapper dataTypeMapper = new DataTypeMapper();
 
     public DMLQueryGenerator(Class<?> clazz) {
         this.table = new TableExtractor(clazz).createTable();
@@ -24,57 +22,36 @@ public class DMLQueryGenerator {
         );
     }
 
-    public String generateSelectQuery(BooleanBuilder booleanBuilder) {
+    public String generateSelectQuery(WhereBuilder whereBuilder) {
         StringBuilder query = new StringBuilder();
         query.append("select ");
         query.append(selectClause());
         query.append(" from ");
         query.append(table.getName());
 
-        if(booleanBuilder.isEmpty()) {
+        if(whereBuilder.isEmpty()) {
             return query.toString();
         }
 
         query.append(" where ");
-        query.append(whereClause(booleanBuilder));
+        query.append(whereBuilder.toClause());
 
         return query.toString();
     }
 
-    public String generateDeleteQuery(BooleanBuilder booleanBuilder) {
+    public String generateDeleteQuery(WhereBuilder whereBuilder) {
         StringBuilder query = new StringBuilder();
         query.append("delete from ");
         query.append(table.getName());
 
-        if(booleanBuilder.isEmpty()) {
+        if(whereBuilder.isEmpty()) {
             return query.toString();
         }
 
         query.append(" where ");
-        query.append(whereClause(booleanBuilder));
+        query.append(whereBuilder.toClause());
 
         return query.toString();
-    }
-
-    private String whereClause(BooleanBuilder booleanBuilder) {
-        StringBuilder stringBuilder = new StringBuilder();
-        boolean firstLine = true;
-
-        for (BooleanExpressionLine line : booleanBuilder.getExpressionLines()) {
-            if (!firstLine) {
-                stringBuilder.append(line.getLogicalOperator().name());
-                stringBuilder.append(" ");
-                firstLine = false;
-            }
-            BooleanExpression expression = line.getExpression();
-            stringBuilder.append(expression.getColumn());
-            stringBuilder.append(" ");
-            stringBuilder.append(expression.getOperator().getSymbol());
-            stringBuilder.append(" ");
-            stringBuilder.append(getValueString(expression.getValue()));
-        }
-
-        return stringBuilder.toString();
     }
 
     private String selectClause() {
@@ -88,17 +65,7 @@ public class DMLQueryGenerator {
     private String valueClause(Columns columns) {
         return columns.getValues()
                 .stream()
-                .map(this::getValueString)
+                .map(ValueUtil::getValueString)
                 .collect(Collectors.joining(", "));
-    }
-
-    public String getValueString(Object value) {
-        if (value == null) {
-            return "null";
-        }
-        if (value instanceof String) {
-            return String.format("'%s'", value);
-        }
-        return value.toString();
     }
 }
